backend/requirements.txt
fastapi
uvicorn
openai-whisper
openai
python-multipart
pydantic
pydantic-settings
python-dotenv
torch
torchvision
torchaudio
ffmpeg-python
